abstractive_summarizer/requirements.txt
streamlit
transformers==4.36.2
torch==2.1.0
newspaper3k
lxml[html_clean]
